add create and edit product functiontality

+++ server/migrations/20210304053455-products.js
@@ -27,7 +27,7 @@ module.exports = {
                 allowNull: false,
             },
             price: {
-                type: Sequelize.INTEGER(),
+                type: Sequelize.DECIMAL(10, 2),
                 allowNull: false,
             },
             description: {
